narrag/nar_rag/requirements.txt
qdrant-client==1.12.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
Pillow>=9.0.0
feedparser>=6.0.0
requests>=2.28.0
google-genai>=1.60.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
python-dateutil>=2.8.2
